Add test for UTF-16 encoding to verify proper Python 2 behavior.
Passing a UTF-16 encoded stream directly to Python 2's csv.reader()
can cause problems because UTF-16 uses NULL bytes in ways that the
reader does not handle. Instead, it's best to recode the stream as
UTF-8 and pass the recoded stream to csv.reader().

Also, changed open() to io.open() in Python 2 code so it behaves
the same as open() in Python 3.